Add support for caching the underlying stream
While php://input is seekable, you can't re-read it a second time. This
is problematic when you want to parse the body for `getParsedBody()`,
but also allow subsequent uses of `getBody()` to work.

To address this, we now allow `Stream` to optionally cache the data it
reads (via `read()` or `getContents()`. Then, once the eof is detected,
we return the cached data on subsequent `getContents()` calls so that
the stream is not read a second time.

Also, update `ServerRequestFactory` so that when it creates a stream for
php://input, it enables the caching mechanism.